+++ Sources/ProjectDescription/TestAction.swift
@@ -2,6 +2,9 @@ import Foundation
 
 /// It represents the test action of a scheme.
 public struct TestAction: Equatable, Codable {
+    /// List of test plans. The first in the list will be the default plan.
+    public let testPlans: [Path]?
+
     /// List of targets to be tested.
     public let targets: [TestableTarget]
 
@@ -32,6 +35,31 @@ public struct TestAction: Equatable, Codable {
     /// Diagnostics options.
     public let diagnosticsOptions: [SchemeDiagnosticsOption]
 
+    private init(testPlans: [Path]?,
+                 targets: [TestableTarget],
+                 arguments: Arguments?,
+                 configurationName: String,
+                 coverage: Bool,
+                 codeCoverageTargets: [TargetReference],
+                 preActions: [ExecutionAction],
+                 postActions: [ExecutionAction],
+                 diagnosticsOptions: [SchemeDiagnosticsOption],
+                 language: String?,
+                 region: String?)
+    {
+        self.testPlans = testPlans
+        self.targets = targets
+        self.arguments = arguments
+        self.configurationName = configurationName
+        self.coverage = coverage
+        self.preActions = preActions
+        self.postActions = postActions
+        self.codeCoverageTargets = codeCoverageTargets
+        self.diagnosticsOptions = diagnosticsOptions
+        self.language = language
+        self.region = region
+    }
+
     /// Initializes a new instance of a test action
     /// - Parameters:
     ///   - targets: List of targets to be tested.
@@ -44,7 +72,7 @@ public struct TestAction: Equatable, Codable {
     ///   - diagnosticsOptions: Diagnostics options.
     ///   - language: Language (e.g. "pl")
     ///   - region: Region (e.g. "PL")
-    public init(targets: [TestableTarget] = [],
+    public init(targets: [TestableTarget],
                 arguments: Arguments? = nil,
                 configurationName: String,
                 coverage: Bool = false,
@@ -55,16 +83,17 @@ public struct TestAction: Equatable, Codable {
                 language: String? = nil,
                 region: String? = nil)
     {
-        self.targets = targets
-        self.arguments = arguments
-        self.configurationName = configurationName
-        self.coverage = coverage
-        self.preActions = preActions
-        self.postActions = postActions
-        self.codeCoverageTargets = codeCoverageTargets
-        self.diagnosticsOptions = diagnosticsOptions
-        self.language = language
-        self.region = region
+        self.init(testPlans: nil,
+                  targets: targets,
+                  arguments: arguments,
+                  configurationName: configurationName,
+                  coverage: coverage,
+                  codeCoverageTargets: codeCoverageTargets,
+                  preActions: preActions,
+                  postActions: postActions,
+                  diagnosticsOptions: diagnosticsOptions,
+                  language: language,
+                  region: region)
     }
 
     /// Initializes a new instance of a test action
@@ -90,7 +119,8 @@ public struct TestAction: Equatable, Codable {
                 language: String? = nil,
                 region: String? = nil)
     {
-        self.init(targets: targets,
+        self.init(testPlans: nil,
+                  targets: targets,
                   arguments: arguments,
                   configurationName: config.name,
                   coverage: coverage,
@@ -100,5 +130,53 @@ public struct TestAction: Equatable, Codable {
                   diagnosticsOptions: diagnosticsOptions,
                   language: language,
                   region: region)
+    }
+
+    /// Initializes a new instance of a test action using test plans
+    /// - Parameters:
+    ///   - testPlans: List of test plans. The first in the list will be the default plan.
+    ///   - config: Configuration that should be used for building the test targets.
+    ///   - preActions: ist of actions to be executed before running the tests.
+    ///   - postActions: List of actions to be executed after running the tests.
+    public static func testPlans(_ testPlans: Path...,
+                                 config: PresetBuildConfiguration = .debug,
+                                 preActions: [ExecutionAction] = [],
+                                 postActions: [ExecutionAction] = []) -> Self
+    {
+        Self(testPlans: testPlans,
+             targets: [],
+             arguments: nil,
+             configurationName: config.name,
+             coverage: false,
+             codeCoverageTargets: [],
+             preActions: preActions,
+             postActions: postActions,
+             diagnosticsOptions: [],
+             language: nil,
+             region: nil)
+    }
+
+    /// Initializes a new instance of a test action using test plans
+    /// - Parameters:
+    ///   - testPlans: List of test plans. The first in the list will be the default plan.
+    ///   - config: Configuration that should be used for building the test targets.
+    ///   - preActions: ist of actions to be executed before running the tests.
+    ///   - postActions: List of actions to be executed after running the tests.
+    public static func testPlans(_ testPlans: Path...,
+                                 configurationName: String,
+                                 preActions: [ExecutionAction] = [],
+                                 postActions: [ExecutionAction] = []) -> Self
+    {
+        Self(testPlans: testPlans,
+             targets: [],
+             arguments: nil,
+             configurationName: configurationName,
+             coverage: false,
+             codeCoverageTargets: [],
+             preActions: preActions,
+             postActions: postActions,
+             diagnosticsOptions: [],
+             language: nil,
+             region: nil)
     }
 }