+++ Sources/TuistGenerator/Generator/SchemeDescriptorsGenerator.swift
@@ -222,6 +222,11 @@ final class SchemeDescriptorsGenerator: SchemeDescriptorsGenerating {
         var preActions: [XCScheme.ExecutionAction] = []
         var postActions: [XCScheme.ExecutionAction] = []
 
+        let testPlans: [XCScheme.TestPlanReference]? = testAction.testPlans?.map {
+            XCScheme.TestPlanReference(reference: "container:\($0.path.relative(to: rootPath))",
+                                       default: $0.isDefault)
+        }
+
         try testAction.targets.forEach { testableTarget in
             guard let reference = try createBuildableReference(targetReference: testableTarget.target,
                                                                graph: graph,
@@ -268,6 +273,7 @@ final class SchemeDescriptorsGenerator: SchemeDescriptorsGenerating {
         return XCScheme.TestAction(buildConfiguration: testAction.configurationName,
                                    macroExpansion: nil,
                                    testables: testables,
+                                   testPlans: testPlans,
                                    preActions: preActions,
                                    postActions: postActions,
                                    shouldUseLaunchSchemeArgsEnv: shouldUseLaunchSchemeArgsEnv,